requirements.txt
  Flask
  Flask-SQLAlchemy
  Flask-Marshmallow
  Marshmallow-SQLAlchemy
  Flask-Login
  python-slugify